Move testsummaries package to testartifact.
